+++ src/types/client.ts
@@ -2,8 +2,9 @@ import { ApolloClient, FetchPolicy } from "apollo-client";
 import { InMemoryCache } from "apollo-cache-inmemory";
 import { ApolloLink } from "apollo-link";
 import { Adapter } from "../adapters/Adapter";
+import { ModelType } from "src";
 
-export interface ClientConfig {
+export interface ClientConfig<T extends ModelType = any> {
   name?: string;
   apollo?: ApolloClient<any>;
   url?: string;