+++ src/types/model.ts
@@ -4,6 +4,12 @@ import { Model } from "../Model";
 
 export type ModelType = typeof Model;
 
+export type ModelConstructor<T, TArgs extends [Attributes] = any> = Function &
+  ModelType & {
+    prototype: T;
+    apply: (this: any, args: TArgs) => any;
+  };
+
 export interface Attributes {
   [field: string]: any;
 }
@@ -16,7 +22,7 @@ export interface FieldCache {
   [key: string]: Fields;
 }
 
-export interface HookCache {
+export interface HookCache<T extends ModelType> {
   [key: string]: Partial<Adapter>;
 }
 
@@ -25,7 +31,7 @@ export interface FieldListCache {
   relationships?: string[];
 }
 
-export type Hooks = Partial<
+export type Hooks<T extends ModelType> = Partial<
   Omit<Adapter, "executeHook"> & {
     $creating(model: Model, data: Attributes): void | Attributes | false;
     $created(model: Model): void;